update listener handler and update e2e test
